refactor: create utils/
The status codes file is now located inside the utils/ directory in the API directory instead of just being present among everything. The sourcing was updated accordingly.
