+++ src/api/dao/user.ts
@@ -1,6 +1,6 @@
 import DbAccess from "../../config/dbAccessFiles/users-model";
 import bcrypt from "bcryptjs";
-import StatusCodes from "../status_codes";
+import StatusCodes from "../utils/status_codes";
 
 class UserDAO {
   /**
